[update] version1.8 分库分表实例二 - 编写基本sql 测试 sharding jdbc 的执行过程

+++ src/main/java/fun/gengzi/codecopy/business/subdata/entity/BussinessTable.java
@@ -51,7 +51,7 @@ public class BussinessTable {
     private Date updatedate;
     //是否删除, 0 否 1 是
     @Column(name = "is_del")
-    private short isDel;
+    private Integer isDel;
     //日志 ID, 可用作日志的查询
     @Column(name = "guid")
     private String guid;
